Refactor hw-ffi crate and add bindings CLI
Modularized the hw-ffi crate by splitting types, errors, and version into separate files. Introduced a new CLI tool for generating Swift and Kotlin bindings, with related Justfile and CI workflow updates. Unified FFI types and error handling, and updated workspace dependencies for consistency. Changed BLE device RSSI type from i16 to i32 for cross-platform compatibility.

+++ crates/ble-transport/src/profile.rs
@@ -39,7 +39,7 @@ impl BleProfile {
 pub struct DeviceInfo {
     pub id: String,
     pub name: Option<String>,
-    pub rssi: Option<i16>,
+    pub rssi: Option<i32>,
     pub services: Vec<Uuid>,
 }
